Test Run 17.1: Sort List

+++ SOS/Utilities/MyException.cpp
@@ -20,19 +20,19 @@ MyException::MyException(const std::stringstream& _exception)
 	code_ = GENERAL;
 }
 
-MyException::MyException(const char * _exception, short _code)
+MyException::MyException(const char * _exception, long _code)
 {
 	exception_ = _exception;
 	code_ = _code;
 }
 
-MyException::MyException(const std::string& _exception, short _code)
+MyException::MyException(const std::string& _exception, long _code)
 {
 	exception_ = _exception;
 	code_ = _code;
 }
 
-MyException::MyException(const std::stringstream& _exception, short _code)
+MyException::MyException(const std::stringstream& _exception, long _code)
 {
 	exception_ = _exception.str();
 	code_ = _code;
